+++ src/scippneutron/file_loading/nxevent_data.py
